+++ packages/faustwp-core/src/hooks/index.ts
@@ -1,0 +1,8 @@
+import { createHooks } from '@wordpress/hooks';
+import { FaustHooks } from './overloads.js';
+
+export type FaustPlugin = {
+  apply: (hooks: FaustHooks) => void;
+};
+
+export const hooks = createHooks();