+++ plugins/faustwp/CHANGELOG.md
@@ -1,4 +1,30 @@
-# FaustWP
+# Faust
+
+## 0.8.1
+
+### Patch Changes
+
+- c016c9f: Update plugin title in changelog
+
+## 0.8.0
+
+### Minor Changes
+
+- b59d6c0: Renamed plugin from `FaustWP` to `Faust`.
+- b59d6c0: Updated settings menu text from _Headless_ to _Faust_.
+
+## 0.7.11
+
+### Patch Changes
+
+- b3c70a4: Prevent WordPress RSS feeds from redirecting to the front-end application.
+
+## 0.7.10
+
+### Patch Changes
+
+- 88ce018: Fix generate endpoint when WordPress is installed within a subdirectory. Props to @kermage for the fix!
+- 0c757a2: Remove unnecessary config from wordpress.org zip.
 
 ## 0.7.9